requirements.txt
mysql-connector-python
sqlalchemy
cloud-sql-python-connector[mysql]
PyMySQL